+++ project/test_master.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 #include <cstddef>
 #include <iostream>
-#include "server.h"
+#include "include/web_server.h"
 
 
 TEST(NULL, check_create) {
